Pizza Application
For my creativity I chose to add an option to make your pizza a deepdish. If you do decide to select deepdish, there is a $10 fee. I also added an option to choose different crusts. The original crust is free, while the gluten free crust adds a $3 fee and the other 3 crusts add just a $2 fee. I also changed the colors of the questions on the application and the order button so it really stands out.